requirements_frozen.txt
albumentations==1.3.1
astroid==2.15.6
-e git+ssh://git@github.com/iankonradjohnson/BatchImageProcessor.git@8b5970ebbd3c32e798f6a6f11c685d581a2b404a#egg=batch_image_processor
black==23.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
contourpy==1.3.2
coverage==7.3.0
cycler==0.12.1
decorator==4.4.2
dill==0.3.7
exceptiongroup==1.1.3
fonttools==4.57.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.0.0
isort==5.12.0
joblib==1.4.2
kiwisolver==1.4.8
lazy-object-proxy==1.9.0
lazy_loader==0.4
matplotlib==3.10.1
mccabe==0.7.0
moviepy==2.1.2
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
proglog==0.1.11
pylint==2.17.5
pyparsing==3.2.3
pytest==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.1
qudida==0.0.4
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.6
typing_extensions==4.7.1
urllib3==2.4.0
wrapt==1.15.0
